Update remake_sentence : rename
