+++ app/ai_cores/remake_cores/__init__.py
@@ -1,3 +1,3 @@
-from .remake import remake_sentence
+from .remake_sentence import remake_sentence
 
 __all__ = ["remake_sentence"]
