+++ app/ai_cores/remake_cores/remake_sentence.py
